% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_point.R
\name{focal_point}
\alias{focal_point}
\title{Define focal_point}
\usage{
focal_point(city = NULL, country = NULL, x = NULL, y = NULL)
}
\arguments{
\item{city}{Name of city appearing in maps::world.cities database}

\item{x}{Longitude of point of interest, provided as alternate to city name}

\item{y}{Latitude of point of interest, provided as alternate to city name}

\item{county}{Name of country of city}
}
\value{
An sf object
}
\description{
Define focal_point
}
\details{
Prompt for x and y coordinate if city name not found
}
\examples{
focal_point(city = "Worcester", country = "USA")
focal_point(city = "Worcester", country = "UK")
focal_point(city = "Worcester", country = "South Africa")
focal_point(x = 15, y = -19.1)
}
