% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ct_cases.R
\name{get_ct_cases}
\alias{get_ct_cases}
\title{Read in cases data from CT DPH daily COVID-19 updates}
\usage{
get_ct_cases(query_date = NULL, URL = "https://portal.ct.gov/Coronavirus")
}
\arguments{
\item{query_date}{User provided date or defaults to NULL (for today)}

\item{URL}{Path to covid-19 page of CT DPH}
}
\value{
Table of county infections
}
\description{
Read in cases data from CT DPH daily COVID-19 updates
}
\details{
CT provides an easier interface to scrape than MA, so the function
is much simpler
}
