% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jhu_daily.R
\name{get_jhu_daily}
\alias{get_jhu_daily}
\title{Read in daily case data from JHU COVID-19 daily reports repository}
\usage{
get_jhu_daily(download_date = NULL, write = FALSE, filepath = NULL)
}
\arguments{
\item{write}{If TRUE, writes output as csv to filepath. Defaults to FALSE}

\item{filepath}{Provide filename and path to write to.}
}
\value{
Long format database of time cases, recoveries, deaths
}
\description{
Read in daily case data from JHU COVID-19 daily reports repository
}
\details{
Downloads daily data from JHU Corona virus repo (master branch).
Data source: https://github.com/CSSEGISandData/COVID-19. If write = TRUE and
fielpath = NULL, a covid19_daily_reports.csv will be written to the current
directory.
}
\examples{
cases <- get_jhu_daily()
cases <- get_jhu_daily(write = TRUE)
}
