% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jhu_ts.R
\name{get_jhu_ts}
\alias{get_jhu_ts}
\title{Read in cases data from JHU COVID-19 case respository}
\usage{
get_jhu_ts(write = FALSE, filepath = NULL)
}
\arguments{
\item{write}{If TRUE, writes output as csv to filepath. Defaults to FALSE}

\item{filepath}{Provide filename and path to write to.}
}
\value{
Long format database of time cases, recoveries, deaths
}
\description{
Read in cases data from JHU COVID-19 case respository
}
\details{
Downloads time series of JHU Corona virus data.
Data source: https://github.com/CSSEGISandData/COVID-19. Code adapted
from: https://github.com/RamiKrispin/coronavirus. If write = TRUE and
fielpath = NULL, a covid19.csv will be written to the current directory
}
\examples{
cases <- get_jhu_ts()
cases <- get_jhu_ts(write = TRUE)
}
