% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jhu_ts.R
\name{get_jhu_ts2}
\alias{get_jhu_ts2}
\title{Read in improved time series data from JHU for the US}
\usage{
get_jhu_ts2(write = FALSE, filepath = NULL)
}
\arguments{
\item{write}{If TRUE, writes output as csv to filepath. Defaults to FALSE}

\item{filepath}{Provide filename and path to write to.}
}
\value{
Long format database of time cases, recoveries, deaths
}
\description{
Read in improved time series data from JHU for the US
}
\details{
Downloads improved time series of JHU Corona virus data for the US,
which is at county-level
Data source: https://github.com/CSSEGISandData/COVID-19.
}
\examples{
cases <- get_jhu_ts2()
cases <- get_jhu_ts2(write = TRUE)
}
