% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_cases.R
\name{us_cases}
\alias{us_cases}
\title{Tidy JHU data for US}
\usage{
us_cases(case_data)
}
\arguments{
\item{case_data}{Tibble of global cases}
}
\value{
Tidy US datasets for state and county level corona virus
}
\description{
Tidy JHU data for US
}
\details{
Cleans downloaded US cases somewhat clumsily
}
