% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dph_cases.R
\name{get_dph_cases}
\alias{get_dph_cases}
\title{Read in cases data from Mass DPH daily docx files}
\usage{
get_dph_cases(mph_date)
}
\arguments{
\item{filepath}{URL of path to daily docx of COVID cases posted to DPH site}
}
\value{
Table of county infections
}
\description{
Read in cases data from Mass DPH daily docx files
}
